fix function definition locations and imports
