+++ pypiper/stage.py
@@ -1,7 +1,6 @@
 """ Conceptualize a pipeline processing phase/stage. """
 
 import copy
-from .utils import translate_stage_name
 
 __author__ = "Vince Reuter"
 __email__ = "vreuter@virginia.edu"
@@ -12,6 +11,7 @@ __all__ = ["Stage"]
 
 CHECKPOINT_EXTENSION = ".checkpoint"
 PIPELINE_CHECKPOINT_DELIMITER = "_"
+STAGE_NAME_SPACE_REPLACEMENT = "-"
 
 
 
@@ -119,3 +119,47 @@ def checkpoint_filename(checkpoint, pipeline_name=None):
         base = "{}{}{}".format(
                 pipeline_name, PIPELINE_CHECKPOINT_DELIMITER, base)
     return base + CHECKPOINT_EXTENSION
+
+
+
+def parse_stage_name(stage):
+    """
+    Determine the name of a stage.
+
+    The stage may be provided already as a name, as a Stage object, or as a
+    callable with __name__ (e.g., function).
+
+    :param stage: Object representing a stage, from which to obtain name.
+    :type stage: str | pypiper.Stage | function
+    :return: Name of putative pipeline Stage.
+    :rtype: str
+    """
+    if isinstance(stage, str):
+        return stage
+    try:
+        return stage.name
+    except AttributeError:
+        try:
+            return stage.__name__
+        except AttributeError:
+            raise TypeError("Unsupported stage type: {}".format(type(stage)))
+
+
+
+def translate_stage_name(stage):
+    """
+    Account for potential variability in stage/phase name definition.
+
+    Since a pipeline author is free to name his/her processing phases/stages
+    as desired, but these choices influence file names, enforce some
+    standardization. Specifically, prohibit potentially problematic spaces.
+
+    :param stage: Pipeline stage, its name, or a representative function.
+    :type stage: str | pypiper.Stage | function
+    :return: Standardized pipeline phase/stage name.
+    :rtype: str
+    """
+    # First ensure that we have text.
+    name = parse_stage_name(stage)
+    # Cast to string to ensure that indexed stages (ints are handled).
+    return str(name).lower().replace(" ", STAGE_NAME_SPACE_REPLACEMENT)
